rename folder, update readme
