+++ src/r1cs_gkr/constraint.rs
@@ -1,4 +1,4 @@
-use crate::circom_gkr::program::SymbolTable;
+use crate::r1cs_gkr::program::SymbolTable;
 use ark_ff::PrimeField;
 use std::cmp::max;
 
@@ -283,10 +283,10 @@ impl<F: PrimeField> From<Term<F>> for (usize, F) {
 
 #[cfg(test)]
 mod tests {
-    use crate::circom_gkr::constraint::{
+    use crate::r1cs_gkr::constraint::{
         move_term_to_slot, Constraint, EquationDirection, Operation, ReducedConstraint, Term,
     };
-    use crate::circom_gkr::program::SymbolTable;
+    use crate::r1cs_gkr::program::SymbolTable;
     use ark_bls12_381::Fr;
 
     #[test]
